http://jira.aplana.com/browse/SBRFACCTAX-8862 Изменения в сервисах скриптов

+++ service/src/main/java/com/aplana/sbrf/taxaccounting/service/impl/PeriodServiceImpl.java
@@ -601,7 +601,7 @@ public class PeriodServiceImpl implements PeriodService {
         List<Months> monthsList = new ArrayList<Months>();
         monthsList.add(null);
 
-        Map<String, RefBookValue> refBookValueMap = dataProvider.getRecordData((long) reportPeriod.getDictTaxPeriodId());
+        Map<String, RefBookValue> refBookValueMap = dataProvider.getRecordData(reportPeriod.getDictTaxPeriodId());
         // Код налогового периода
         String code = refBookValueMap.get("CODE").getStringValue();
